requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10

channels~=2.4.0